Discusses trace propagation concerns with trace reporting (#42)
This formalizes a growing problem where tracing systems are used in a mesh
environment, instrumented with B3, and proxying the trace data endpoint.

This was noticed before in Zipkin, but the problem was rare. Now that it is
a routine concern, it is worth informing others using B3 format.

Thanks @basvanbeek for the analysis and @anuraaga for input leading to
this.